Change structure and added helper function
